+++ app/src/main/java/umc/study/umc_7th/main/BottomNavigationBar.kt
@@ -1,4 +1,4 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.main
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -32,6 +32,9 @@ import androidx.compose.ui.text.style.LineBreak
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import umc.study.umc_7th.Content
+import umc.study.umc_7th.R
+import umc.study.umc_7th.getTestMusicContentList
 
 enum class NavigationDestination(
     val expression: String,
