+++ app/src/main/java/umc/study/umc_7th/main/album/AlbumDetailsPage.kt
@@ -1,23 +1,25 @@
-package umc.study.umc_7th.locker
+package umc.study.umc_7th.main.album
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
 
 @Composable
-fun StorageFilePage() {
+fun AlbumDetailsPage() {
     // TODO: 구현
     Box(
         contentAlignment = Alignment.Center,
-        modifier = Modifier.fillMaxSize()
+        modifier = Modifier.fillMaxWidth().height(100.dp)
     ) {
         Text(
-            text = "파일이 없습니다.",
+            text = "상세 정보가 없습니다.",
             style = TextStyle(
                 color = Color.Black.copy(alpha = 0.5f),
             )
