+++ app/src/main/java/umc/study/umc_7th/main/album/IncludedContentsPage.kt
@@ -1,4 +1,4 @@
-package umc.study.umc_7th.album
+package umc.study.umc_7th.main.album
 
 import android.os.Build
 import androidx.annotation.RequiresApi
