+++ app/src/main/java/umc/study/umc_7th/main/locker/StorageFilePage.kt
@@ -1,25 +1,23 @@
-package umc.study.umc_7th.album
+package umc.study.umc_7th.main.locker
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.dp
 
 @Composable
-fun AlbumDetailsPage() {
+fun StorageFilePage() {
     // TODO: 구현
     Box(
         contentAlignment = Alignment.Center,
-        modifier = Modifier.fillMaxWidth().height(100.dp)
+        modifier = Modifier.fillMaxSize()
     ) {
         Text(
-            text = "상세 정보가 없습니다.",
+            text = "파일이 없습니다.",
             style = TextStyle(
                 color = Color.Black.copy(alpha = 0.5f),
             )
